pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3", "wheel", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mia*"]

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.setuptools-git-versioning]
enabled = true

[project]
name = "miarchive"
dynamic = ["version"]
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
    # Archival
    "undetected-geckodriver-lw",
    "selenium-wire-lw[uf]>=2.1.0",

    # Misc. processing
    "beautifulsoup4==4.13.3",

    # Data storage
    "psycopg==3.2.9",

    # Server
    "aiohttp==3.12.13",
    "aiohttp-jinja2==1.6",
    "aiohttp_cors==0.8.1",

    # Deserialisation
    "msgspec>=0.19.0",

    # Logging
    "loguru>=0.7.3",

    # Misc utilities
    "requests",
    "xvfbwrapper==0.2.13",
]
readme = "README-pypi.md"

authors = [
    { name = "LunarWatcher", email = "oliviawolfie@pm.me" },
]

description = "Self-hosted internet archive"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",

    "Topic :: System :: Archiving",
    "Topic :: Utilities",

    "Framework :: aiohttp",

    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: System Administrators",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",

    "Operating System :: POSIX :: Linux",
]
keywords = [
    "archival", "web archiving", "internet archiving", "web", "browser", "download",
    "preservation", 
]

[project.scripts]
mia = "mia:cli_main.main"

[project.urls]
Homepage = "https://codeberg.org/LunarWatcher/MIArchive"
Documentation = "https://codeberg.org/LunarWatcher/MIArchive"
Repository = "https://codeberg.org/LunarWatcher/MIArchive.git"
Issues = "https://codeberg.org/LunarWatcher/MIArchive/issues"
Changelog = "https://codeberg.org/LunarWatcher/MIArchive/src/branch/master/CHANGELOG.md"
